+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to **Enumeration** will be documented in this file. The proj
 
 ## [Unreleased]
 
+## [1.2.0] - 2026-04-06
+
 ### Added
 
 - Support for `partial class` declarations in addition to `partial record`.
@@ -36,7 +38,8 @@ All notable changes to **Enumeration** will be documented in this file. The proj
 
 - Initial release of `Enumeration` source code generator.
 
-[unreleased]: https://github.com/linkdotnet/Enumeration/compare/1.1.0...HEAD
+[unreleased]: https://github.com/linkdotnet/Enumeration/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/linkdotnet/Enumeration/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/linkdotnet/Enumeration/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/linkdotnet/Enumeration/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/linkdotnet/Enumeration/compare/8d85f242bf1652588c7b544b297c6734e1044e3d...1.0.0